requirements.txt
fastapi
uvicorn[standard]
chromadb
python-dotenv
requests
langchain
openai
langchain-community
tiktoken